Put the reshape utils in gpu/copy.h
