+++ mlx/backend/cuda/conv.cpp
@@ -4,7 +4,6 @@
 #include "mlx/backend/cuda/device/config.h"
 #include "mlx/backend/cuda/lru_cache.h"
 #include "mlx/backend/gpu/copy.h"
-#include "mlx/backend/gpu/reshape.h"
 #include "mlx/dtype_utils.h"
 #include "mlx/primitives.h"
