+++ mlx/backend/gpu/CMakeLists.txt
@@ -2,5 +2,4 @@ target_sources(
   mlx
   PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/copy.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/primitives.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/reshape.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/slicing.cpp)
